Refactor: wishList 관련 함수 product에서 user로 이동
