+++ server/controllers/index.js
@@ -1,9 +1,11 @@
 const authController = require("./authController");
 const productController = require("./productController");
 const chatController = require("./chatController");
+const userController = require("./userController");
 
 module.exports = {
   authController: authController,
   productController: productController,
   chatController: chatController,
+  userController: userController,
 };
