+++ server/routes/index.js
@@ -1,6 +1,7 @@
 const auth = require("./auth");
 const product = require("./product");
 const chat = require("./chat");
+const user = require("./user");
 
 /** 라우터 통합 실행 함수
  * @description 모든 api 라우팅
@@ -11,6 +12,7 @@ const routerLoader = () => {
   auth(app);
   product(app);
   chat(app);
+  user(app);
   return app;
 };
